chore(connect): rework gen-reftx util to ts
